refactor(enums): extract HasPrefixedWireType trait
FieldType and ColumnType carried byte-identical wireType/localType prefix logic;
move it to a shared enum trait, leaving each enum only its Prefix constant.
