+++ src/Tables/Enums/ColumnType.php
@@ -3,8 +3,12 @@ declare(strict_types=1);
 
 namespace Lattice\Lattice\Tables\Enums;
 
+use Lattice\Lattice\Core\Enums\Concerns\HasPrefixedWireType;
+
 enum ColumnType: string
 {
+    use HasPrefixedWireType;
+
     private const string Prefix = 'column.';
 
     case Text = 'column.text';
@@ -15,20 +19,4 @@ enum ColumnType: string
     case Badge = 'column.badge';
     case Icon = 'column.icon';
     case Image = 'column.image';
-
-    public static function wireType(self|string $type): string
-    {
-        if ($type instanceof self) {
-            return $type->value;
-        }
-
-        return str_starts_with($type, self::Prefix) ? $type : self::Prefix.$type;
-    }
-
-    public static function localType(self|string $type): string
-    {
-        $value = $type instanceof self ? $type->value : $type;
-
-        return str_starts_with($value, self::Prefix) ? substr($value, strlen(self::Prefix)) : $value;
-    }
 }
